Tried to Fix Color
Tried to actually get rid of the black planets, but nothing seems to work. Also, I don't know how to lower the cap on the speed to where there is less of a chance of lag being an issue. I've been trying for a bit but here is what I have done so far.